core_simulation/pyproject.toml
[project]
name = "prompt-eng"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask==3.0.2",
    "flask-socketio==5.3.6",
    "numpy==1.26.4",
    "python-dotenv==1.0.1",
    "python-engineio==4.9.1",
    "python-socketio==5.11.1",
    "uvicorn==0.27.1",
    "websockets==12.0",
]
